+++ crates/warpgate/src/loader.rs
@@ -24,7 +24,7 @@ pub type OfflineChecker = Arc<fn() -> bool>;
 #[derive(Clone)]
 pub struct PluginLoader {
     /// Instance of our HTTP client.
-    http_client: OnceCell<reqwest::Client>,
+    http_client: OnceCell<Arc<reqwest::Client>>,
 
     /// Options to pass to the HTTP client.
     http_options: HttpOptions,
@@ -60,9 +60,9 @@ impl PluginLoader {
     }
 
     /// Return the HTTP client, or create it if it does not exist.
-    pub fn get_client(&self) -> miette::Result<&reqwest::Client> {
+    pub fn get_client(&self) -> miette::Result<&Arc<reqwest::Client>> {
         self.http_client
-            .get_or_try_init(|| create_http_client_with_options(&self.http_options))
+            .get_or_try_init(|| create_http_client_with_options(&self.http_options).map(Arc::new))
     }
 
     /// Load a plugin using the provided locator. File system plugins are loaded directly,