+++ plugins/wasm-test/src/lib.rs
@@ -10,6 +10,7 @@ extern "ExtismHost" {
     fn from_virtual_path(path: String) -> String;
     fn get_env_var(name: String) -> String;
     fn host_log(input: Json<HostLogInput>);
+    fn send_request(input: Json<SendRequestInput>) -> Json<SendRequestOutput>;
     fn set_env_var(name: String, value: String);
     fn to_virtual_path(path: String) -> String;
 }
@@ -51,6 +52,10 @@ pub fn testing_macros(_: ()) -> FnResult<()> {
     let _ = exec_command!(raw, "git");
     let _ = exec_command!(raw, "git", args);
     let _ = exec_command!(raw, "git", ["a", "b", "c"]);
+
+    // Requests
+    send_request!("https://some/url");
+    send_request!(input, SendRequestInput::new("https://some/url"));
 
     // Env vars
     let name = "VAR";
@@ -201,8 +206,7 @@ struct NodeDistVersion {
 #[plugin_fn]
 pub fn load_versions(Json(_): Json<LoadVersionsInput>) -> FnResult<Json<LoadVersionsOutput>> {
     let mut output = LoadVersionsOutput::default();
-    let response: Vec<NodeDistVersion> =
-        fetch_url_with_cache("https://nodejs.org/dist/index.json")?;
+    let response: Vec<NodeDistVersion> = fetch_json("https://nodejs.org/dist/index.json")?;
 
     for (index, item) in response.iter().enumerate() {
         let version = Version::parse(&item.version[1..])?;